+++ apps/vis_app/static/vis_app/ts/vis/types.ts
@@ -1,5 +1,5 @@
 /**
- * Type definitions for Sigma Editor
+ * Type definitions for Vis Editor
  */
 
 export interface Dataset {